requirements.txt
langchain-google-genai
pillow
langchain
langchain_community
langchain_openai
python-dotenv
streamlit
langchain_experimental
sentence-transformers
langchain_chroma
langchainhub
unstructured